Reorganize Hibernate Search management config for a Standalone mapper
