+++ extensions/hibernate-search-standalone-elasticsearch/runtime/src/main/java/io/quarkus/hibernate/search/standalone/elasticsearch/runtime/HibernateSearchStandaloneBuildTimeConfig.java
@@ -69,6 +69,12 @@ public interface HibernateSearchStandaloneBuildTimeConfig {
     Optional<String> backgroundFailureHandler();
 
     /**
+     * Management interface.
+     */
+    @ConfigDocSection
+    HibernateSearchStandaloneBuildTimeConfigManagement management();
+
+    /**
      * Configuration related to the mapping.
      */
     MappingConfig mapping();
